Update Jenkinsfile
tfgf

+++ Jenkinsfile
@@ -1,6 +1,10 @@
 pipeline {
     agent any
     
+    environment {
+        CONTAINER_CREATED = false
+    }
+    
     stages {
         stage('Build Docker Image') {
             steps {
@@ -10,27 +14,31 @@ pipeline {
             }
         }
         stage('Run Docker Container') {
-    steps {
-        script {
-            // Remove previous container if it exists
-            sh "docker rm -f my-docker-container-${env.BUILD_NUMBER} || true"
+            steps {
+                script {
+                    // Remove previous container if it exists
+                    sh "docker rm -f my-docker-container-${env.BUILD_NUMBER} || true"
 
-            // Run the new container
-            sh "docker run -d -p 8081:80 --name my-docker-container-${env.BUILD_NUMBER} my-docker-image:${env.BUILD_NUMBER}"
+                    // Run the new container
+                    def exitCode = sh script: "docker run -d -p 8081:80 --name my-docker-container-${env.BUILD_NUMBER} my-docker-image:${env.BUILD_NUMBER}", returnStatus: true
+                    if (exitCode == 0) {
+                        env.CONTAINER_CREATED = true
+                    }
+                }
+            }
         }
     }
-}
-
-    }
-
+    
     post {
         always {
             // Clean up any resources here
             script {
-                sh "docker rm -f my-docker-container-${env.BUILD_NUMBER} || true"
+                if (env.CONTAINER_CREATED) {
+                    sh "docker rm -f my-docker-container-${env.BUILD_NUMBER} || true"
+                }
                 docker.image("my-docker-image:${env.BUILD_NUMBER}").remove()
             }
         }
     }    
-
 }
+
